core: move all models to namespace
