+++ BlasII.CustomSkins/CustomSkins.cs
@@ -2,6 +2,7 @@
 using BlasII.CustomSkins.Exporters;
 using BlasII.CustomSkins.Finders;
 using BlasII.CustomSkins.Importers;
+using BlasII.CustomSkins.Models;
 using BlasII.ModdingAPI;
 using BlasII.ModdingAPI.Helpers;
 using Il2CppTGK.Game;
