+++ BlasII.CustomSkins/Exporters/FakeExporter.cs
@@ -1,4 +1,5 @@
-﻿using BlasII.ModdingAPI;
+﻿using BlasII.CustomSkins.Models;
+using BlasII.ModdingAPI;
 using System.Collections;
 using System.Linq;
