+++ BlasII.CustomSkins/Extensions/DictionaryExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using BlasII.CustomSkins.Models;
 using UnityEngine;
 
 namespace BlasII.CustomSkins.Extensions;
