+++ BlasII.CustomSkins/Importers/IImporter.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections;
+using BlasII.CustomSkins.Models;
 
 namespace BlasII.CustomSkins.Importers;
